Validering av tekst-søk på sak-title og between dates på saksdato.
