+++ api/KS.FiksProtokollValidator.Tests/KS.FiksProtokollValidator.Tests.csproj
@@ -37,6 +37,18 @@
     <Content Include="TestData\ValidatorTests\NySokN1\sokeresultatMinimum2.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Content>
+    <Content Include="TestData\ValidatorTests\NySokN4\sok.xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestData\ValidatorTests\NySokN4\sokeresultatMinimum.xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestData\ValidatorTests\NySokN4\sokeresultatMinimumInvalid.xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
+    <Content Include="TestData\ValidatorTests\NySokN4\sokeresultatMinimumSaksdatoFeil.xml">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </Content>
     <None Remove="TestData\Requests\NySokN1\sokeresultatMinimum2.xml" />
     <None Remove="TestData\Requests\NySokN1\sokeresultatMinimum.xml" />
   </ItemGroup>
@@ -74,10 +86,6 @@
     <None Update="TestData\Responses\svar_paa_soek.xml">
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="TestData\ValidatorTests" />
   </ItemGroup>
 
 </Project>
