+++ api/KS.FiksProtokollValidator.Tests/TestData/ValidatorTests/NySokN4/sok.xml
@@ -1,0 +1,20 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<sok xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://www.ks.no/standarder/fiks/arkiv/sok/v1">
+	<system>Fagsystem X</system>
+	<meldingId>09f6cf81-cb45-4abc-9065-68f5abb3d866</meldingId>
+	<tidspunkt>2021-03-23T10:28:33.1952112+01:00</tidspunkt>
+	<take>100</take>
+	<skip>0</skip>
+	<respons>saksmappe</respons>
+	<parameter>
+		<felt>sak.saksdato</felt>
+		<operator>between</operator>
+		<parameterverdier>
+			<datevalues>
+				<value>2021-11-01T00:00:00.0</value>
+				<value>2021-12-31T23:59:59.0</value>
+			</datevalues>
+		</parameterverdier>
+	</parameter>
+	<responsType>minimum</responsType>
+</sok>
