+++ api/KS.FiksProtokollValidator.WebAPI/Validation/AbstractValidator.cs
@@ -7,18 +7,6 @@ namespace KS.FiksProtokollValidator.WebAPI.Validation
 {
     public abstract class AbstractValidator
     {
-        protected static List<DateTime> GetDateResults(SokeresultatMinimum sokResponse, SokFelt parameterFelt)
-        {
-            switch (parameterFelt)
-            {
-                case SokFelt.SakPeriodSaksdato:
-                    return sokResponse.ResultatListe.Select(r => r.Saksmappe.Saksdato).ToList();
-                
-                default:
-                    return new List<DateTime>();
-            }
-        }
-
         protected static void ValidateBetweenDates(DateTime value, DateTime dateFrom, DateTime dateTo,
             ICollection<string> validationErrors)
         {
