--- api/KS.FiksProtokollValidator.WebAPI/Validation/ParseHelpers/ISokeresultatParseHelper.cs
@@ -1,7 +1,0 @@
-namespace KS.FiksProtokollValidator.WebAPI.Validation.ParseHelpers
-{
-    public interface ISokeresultatParseHelper
-    {
-        
-    }
-}
