--- api/KS.FiksProtokollValidator.WebAPI/Validation/ParseHelpers/SokeresultatMinimumParseHelper.cs
@@ -1,7 +1,0 @@
-namespace KS.FiksProtokollValidator.WebAPI.Validation.ParseHelpers
-{
-    public class SokeresultatMinimumParseHelper
-    {
-        
-    }
-}
